requirements.txt
africastalking==1.2.9
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
Django==5.1.3
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.0
fpdf==1.7.2
idna==3.10
phonenumbers==9.0.5
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dotenv==1.1.0
requests==2.32.3
schema==0.7.7
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
tzdata==2024.2
urllib3==2.3.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
gunicorn
dj-database-url
